backend/requirements.txt
fastapi
fastapi_sqlalchemy
uvicorn
pydantic
passlib
bcrypt
python-jose
python-multipart
pyodbc
requests